Fix code scanning alert no. 7: Implicit narrowing conversion in compound assignment

+++ testcore/test/java/hgtest/indexing/ManyToManyIndexerTests.java
@@ -161,7 +161,7 @@ public class ManyToManyIndexerTests extends HGTestBase
         }
         result.close();
         result = index.scanKeys();
-        int totalValues2 = 0;
+        long totalValues2 = 0;
         while (result.hasNext())
         {
             long forkey = index.count(result.next());
